Fixed REST API

+++ src/main/java/pl/football/league/API/coach/CoachController.java
@@ -1,17 +1,17 @@
 package pl.football.league.API.coach;
 
 import pl.football.league.API.team.TeamService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.annotation.Resource;
 import java.util.List;
 
 @RestController
 public class CoachController {
 
-    @Autowired
+    @Resource
     private CoachService coachService;
-    @Autowired
+    @Resource
     private TeamService teamService;
 
     @RequestMapping("/coaches")
@@ -29,19 +29,21 @@ public class CoachController {
         return coachService.findCoach(name, surname);
     }
 
-    @RequestMapping(method =  RequestMethod.POST, value="/coaches/add")
+    @PostMapping("/coaches/add")
     public void addCoach(@RequestBody Coach coach){
         coachService.addCoach(coach);
     }
 
-    @RequestMapping(method =  RequestMethod.PUT, value="/coaches/update")
-    public void updateCoach(@RequestBody Coach coach){
-        coachService.updateCoach(coach);
+    @PutMapping("/coaches/{id}/update")
+    public void updateCoach(@RequestBody Coach coach,@PathVariable long id){
+        if(coach.getCoachID() == id)
+            coachService.updateCoach(coach);
     }
 
-    @RequestMapping(method = RequestMethod.DELETE, value = "/coaches/delete")
-    public void deleteCoach(@RequestBody Coach coach){
-        if(teamService.getTeamByCoach(coach) == null)
-            coachService.deleteCoach(coach);
+    @DeleteMapping("/coaches/{id}/delete")
+    public void deleteCoach(@RequestBody Coach coach, @PathVariable long id){
+        if(coach.getCoachID() == id)
+            if(teamService.getTeamByCoach(coach) == null)
+                coachService.deleteCoach(coach);
     }
 }
